Modify MathchingChallenge to take encoded pairs
Having issues with TypeScript passing in the wrong classes due to type definitions.
Until I get round to a more extensive rewrite, this modification works well because
we can just pass strings and decode them in the constructor.
For the time being, I'm going to keep the MatchingChallenge constructor as the only
point of entry to this module, so the decoding strategy will work as long as that
remains the case.